recursion practive
took a good minute to understand this. i just stared at the problem untill it clicked. my mind was stuck on the fact that there was no loop and then it occured to me that the loop is that at the end the function calls its self wile adusting a value. as the update value goes through the function again it passes through the else if statement and if it gets back to the function at the bottom it calls its self again until the else if stops it.