Create array practice
this one took some time. I eventually came to this program. i think i can refactor a bit using arrow functions and possibly assigning "rangeArray.push(i);" to a variable. i can also add a way to confirm that when the start is higher then the stop that if the third parameter is added needs to be a negative integer.